reorganizing the structure of directories (global change)
The old structure of directories was something new, whereas
the essential links for compatibility with the standard FHS,
were kept.

Now the structure is more much like the FHS standard, given
the following reasons:

- less options
- less hardcode
- less symlinks
- compatibility for static packages with other Unix-like systems

The file containing the new structure is called "_hierarchyfs".
You can see the tree of directories on: "archive/_hierarchyfs.txt".