bootstrap.d/0/03-binutils: clarify host configuration
When used with --target, this creates a cross-architecture
executable that creates files for $TARGET but runs on $HOST.

We must ensure that the configuration of the host is
that of the given host. Otherwise, host relies using
the same target triplet as the building system.

AR="ar" prevents binutils from compiling with ${HOST}-ar
as they are provided by this his package and therefore not
installed yet.