stage 1: disable ping and wget in busybox
I don't know how to set properly the internet connection or how to get
a shared connection for busybox.  The main purpose would be to download
the sources from the chroot.  This is a not big issue anyway, because
the sources are available for download from the mirrors of Dragora.